requirements.txt
influxdb-client==1.16.0
speedtest-cli==2.1.3
